display: Add QTI gralloc4 configuration and replace USE_GRALLOC1 cflag
This is based on:
https://review.lineageos.org/c/LineageOS/android_vendor_lineage/+/358586/2

However, instead of configuring a separate globalVar, let's make use of the
fact that legacy platforms prior sdm845 use GRALLOC1 and sdm845 and above
use GRALLOC4 as they could use the QSSI headers without problems with
GRALLOC4=true. Hence, let's instead make use of TARGET_USES_GRALLOC4 to
indicate whether we change the header arrangement or not.

Change-Id: I6481c9094c9545bfb181675c5257d1e7b014399c
Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>